changed data type in schema from 'integer' to 'bigint' to include numbers big enough to record milage, price, and years

+++ db/migrate/20180404014957_create_vehicles.rb
@@ -2,18 +2,18 @@ class CreateVehicles < ActiveRecord::Migration[5.1]
   def change
     create_table :vehicles do |t|
       t.string :name
-      t.integer :model_year
+      t.bigint :model_year
       t.string :make
       t.string :model
       t.string :trim_level
       t.string :body_style
       t.string :engine_type
-      t.integer :milage
+      t.bigint :milage
       t.string :vin
       t.text :notes
       t.date :bought_date
-      t.integer :bought_milage
-      t.integer :bought_price
+      t.bigint :bought_milage
+      t.bigint :bought_price
 
       t.timestamps
     end
